legacy_backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Neo4j driver
neo4j==5.24.0

# OpenAI SDK
openai==1.51.0

# Code parsing
tree-sitter==0.23.1
tree-sitter-python==0.23.2

# Settings management
pydantic-settings==2.5.2
pydantic==2.9.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# Utilities
python-dotenv==1.0.1

# Multipart form support (for ZIP upload endpoint)
python-multipart==0.0.12

# Git operations (for GitHub URL ingestion)
gitpython==3.1.43
